requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements/requirements.in
#
asgiref==3.3.1            # via django
attrs==20.3.0             # via pytest
django==3.1.4             # via -r requirements/requirements.in, djangorestframework
djangorestframework==3.12.2  # via -r requirements/requirements.in
importlib-metadata==3.3.0  # via pluggy, pytest
iniconfig==1.1.1          # via pytest
packaging==20.8           # via pytest
pluggy==0.13.1            # via pytest
py==1.10.0                # via pytest
pyparsing==2.4.7          # via packaging
pytest==6.2.1             # via -r requirements/requirements.in
pytz==2020.5              # via django
sqlparse==0.4.1           # via django
toml==0.10.2              # via pytest
typing-extensions==3.7.4.3  # via importlib-metadata
zipp==3.4.0               # via importlib-metadata
